+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 
 ### Bug Fixes
 - Fix after-login redirects (the "?next=" query parameter)
+- Update Check.status field when user edits timeout & grace settings
 
 
 ## 1.3.0 - 2018-11-21